requirements.txt
-e .
aws_cdk.aws_kinesis
aws_cdk.aws_ec2
aws_cdk.aws_kinesis
aws_cdk.aws_ssm
aws_cdk.aws_lambda
aws_cdk.aws_logs